+++ src/gmod-database-webserver/index.js
@@ -1,6 +1,5 @@
 const express = require('express');
 const bodyParser = require('body-parser');
-const dbCon = require('./rethinkdb');
 
 const apiRouter = require('./routes/api/v1')
 
@@ -12,7 +11,5 @@ app.use(bodyParser.json())
 const webServer = app.listen(80)
 
 process.on('SIGTERM', () => {
-  console.log('Stopping Webserver')
-  dbCon.close();
   webServer.close();
 })
